+++ core/src/main/java/com/schurke/game/core/GameConfig.java
@@ -1,4 +1,4 @@
-package com.schurke.game;
+package com.schurke.game.core;
 
 public class GameConfig {
     private static boolean unlimitedAmmo = false;
