+++ core/src/main/java/com/schurke/game/entities/Enemy.java
@@ -1,4 +1,4 @@
-package com.schurke.game;
+package com.schurke.game.entities;
 
 import java.util.ArrayList;
