+++ core/src/main/java/com/schurke/game/input/AimRenderer.java
@@ -1,10 +1,11 @@
-package com.schurke.game;
+package com.schurke.game.input;
 
 import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.graphics.OrthographicCamera;
 import com.badlogic.gdx.graphics.glutils.ShapeRenderer;
 import com.badlogic.gdx.math.Vector2;
 import com.badlogic.gdx.math.Vector3;
+import com.schurke.game.entities.Player;
 
 public class AimRenderer {
     private final OrthographicCamera camera;
