+++ core/src/main/java/com/schurke/game/map/TileMap.java
@@ -1,4 +1,4 @@
-package com.schurke.game;
+package com.schurke.game.map;
 import com.badlogic.gdx.graphics.Texture;
 import com.badlogic.gdx.graphics.g2d.SpriteBatch;
 import com.badlogic.gdx.Gdx;
