+++ core/src/main/java/com/schurke/game/screens/GameOverScreen.java
@@ -1,4 +1,4 @@
-package com.schurke.game;
+package com.schurke.game.screens;
 
 import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.Screen;
@@ -14,6 +14,7 @@ import com.badlogic.gdx.scenes.scene2d.ui.TextButton.TextButtonStyle;
 import com.badlogic.gdx.scenes.scene2d.utils.ChangeListener;
 import com.badlogic.gdx.graphics.Color;
 import com.badlogic.gdx.utils.viewport.ScreenViewport;
+import com.schurke.game.Main;
 import com.badlogic.gdx.scenes.scene2d.ui.Label.LabelStyle;
 
 public class GameOverScreen implements Screen {
@@ -28,30 +29,30 @@ public class GameOverScreen implements Screen {
             this.game = game;
             this.batch = new SpriteBatch();
             stage = new Stage(new ScreenViewport());
-            
+
             // Create fonts
             font = new BitmapFont();
             font.getData().setScale(2.0f);
             titleFont = new BitmapFont();
             titleFont.getData().setScale(3.0f);
             titleFont.setColor(Color.RED);
-            
+
             // Create button style
             TextButtonStyle textButtonStyle = new TextButtonStyle();
             textButtonStyle.font = font;
             textButtonStyle.fontColor = Color.WHITE;
             textButtonStyle.downFontColor = Color.LIGHT_GRAY;
-            
+
             // Create label style for title
             LabelStyle labelStyle = new LabelStyle(titleFont, Color.RED);
-            
+
             // Create table for layout
             Table table = new Table();
             table.setFillParent(true);
-            
+
             // Create title label
             Label titleLabel = new Label("Game Over!", labelStyle);
-            
+
             // Create restart button
             TextButton restartButton = new TextButton("Restart Game", textButtonStyle);
             restartButton.addListener(new ChangeListener() {
@@ -65,7 +66,7 @@ public class GameOverScreen implements Screen {
                     }
                 }
             });
-            
+
             // Create exit button
             TextButton exitButton = new TextButton("Exit Game", textButtonStyle);
             exitButton.addListener(new ChangeListener() {
@@ -79,15 +80,15 @@ public class GameOverScreen implements Screen {
                     }
                 }
             });
-            
+
             // Add widgets to table with spacing
             table.add(titleLabel).padBottom(50).row();
             table.add(restartButton).pad(10).row();
             table.add(exitButton).pad(10);
-            
+
             stage.addActor(table);
             Gdx.input.setInputProcessor(stage);
-            
+
         } catch (Exception e) {
             Gdx.app.error("GameOverScreen", "Error initializing game over screen", e);
             throw e; // Rethrow to ensure the error is not silently swallowed
@@ -99,10 +100,10 @@ public class GameOverScreen implements Screen {
         try {
             Gdx.gl.glClearColor(0.15f, 0.15f, 0.2f, 1f);
             Gdx.gl.glClear(GL20.GL_COLOR_BUFFER_BIT);
-            
+
             stage.act(Math.min(delta, 1/30f)); // Cap delta time to prevent huge jumps
             stage.draw();
-            
+
         } catch (Exception e) {
             Gdx.app.error("GameOverScreen", "Error in render", e);
         }
