+++ core/src/main/java/com/schurke/game/weapons/Weapon.java
@@ -1,8 +1,9 @@
-package com.schurke.game;
+package com.schurke.game.weapons;
 
 import java.util.List;
 
 import com.badlogic.gdx.math.Vector2;
+import com.schurke.game.combat.Bullet;
 
 public interface Weapon {
     List<Bullet> shoot(Vector2 position, Vector2 direction);
